+++ libraries/bio/src/utils/generator.ts
@@ -1,7 +1,11 @@
-import * as grok from 'datagrok-api/grok';
 import * as DG from 'datagrok-api/dg';
-import * as ui from 'datagrok-api/ui';
-import {ALPHABET, NOTATION, TAGS} from './macromolecule';
+import {ALPHABET, getAlphabet, NOTATION, TAGS} from './macromolecule';
+
+import wu from 'wu';
+import {GapOriginals, SeqHandler} from './seq-handler';
+import {StringListSeqSplitted} from './macromolecule/utils';
+import {IMonomerLib} from '../types/index';
+import {PolymerTypes} from './const';
 
 export function generateManySequences(): DG.Column[] {
   const columns: DG.Column[] = [];
@@ -15,7 +19,6 @@ export function generateManySequences(): DG.Column[] {
 
 /** Generates the column 'MSA' with sequences length of order 10^6 and the 'Activity' float column. */
 export function generateLongSequence(length: number = 10 ** 5): DG.Column[] {
-  const columns: DG.Column[] = [];
   const longSequence =
     `meI/hHis/Aca/N//dE/Thr_PO3H2/Aca/D-Tyr_Et/Tyr_ab-dehydroMe`.repeat(Math.ceil(length / 10)).slice(0, -1);
   const msaCol = DG.Column.fromList(DG.COLUMN_TYPE.STRING, 'MSA', new Array(10 ** 2).fill(longSequence));
@@ -25,7 +28,36 @@ export function generateLongSequence(length: number = 10 ** 5): DG.Column[] {
   msaCol.setTag(TAGS.alphabet, ALPHABET.UN);
   msaCol.setTag(TAGS.alphabetIsMultichar, 'true');
 
+  const columns: DG.Column[] = [];
   columns.push(msaCol);
   columns.push(DG.Column.fromList(DG.COLUMN_TYPE.FLOAT, 'Activity', new Array(10 ** 2).fill(7.30751)));
   return columns;
+}
+
+export function generateLongSequence2(
+  notation: NOTATION = NOTATION.SEPARATOR, alphabet: ALPHABET = ALPHABET.PT,
+  separator: string | undefined = (notation === NOTATION.SEPARATOR ? '-' : undefined),
+  monomerLib: IMonomerLib | undefined = undefined,
+  colName: string = 'seq', rowCount: number = 100, seqLength: number = 10 ** 6
+): DG.Column {
+  const alphabetSet: string[] = alphabet === ALPHABET.UN ?
+    monomerLib?.getMonomerSymbolsByType(PolymerTypes.PEPTIDE) ?? [] : Array.from(getAlphabet(alphabet));
+  const alphabetSize = alphabetSet.length;
+
+  const col = DG.Column.fromType(DG.COLUMN_TYPE.STRING, colName, rowCount);
+  col.semType = DG.SEMTYPE.MACROMOLECULE;
+  col.setTag(DG.TAGS.UNITS, notation);
+  col.setTag(TAGS.alphabet, alphabet);
+  if (notation == NOTATION.SEPARATOR) col.setTag(TAGS.separator, separator!);
+
+  const sh = SeqHandler.forColumn(col);
+  for (let rowI = 0; rowI < rowCount; rowI++) {
+    const seqMList: string[] = wu.count(0).take(seqLength)
+      .map((i) => { return alphabetSet[Math.floor(Math.random() * alphabetSize)]; })
+      .toArray();
+    const seq = sh.joiner(new StringListSeqSplitted(seqMList, GapOriginals[notation]));
+    col.set(rowI, seq);
+  }
+
+  return col;
 }